Add files via upload
This project is a real-time **Face Recognition and Emotion Detection System** built using Python, OpenCV, and the `FER` library. It captures live video from the webcam, detects faces, and identifies emotions such as happiness, sadness, and anger. The system uses a confidence threshold to ensure accurate predictions and smooths results by averaging emotions detected over multiple frames. It provides a visual output with bounding boxes around detected faces and displays the identified emotion on the screen. This project is ideal for exploring computer vision and emotion analysis in real-time applications.